deepseek/requirements.txt
# requirements.txt
torch>=1.9.0
torchvision>=0.10.0
mlflow>=1.20.0
numpy>=1.21.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
librosa>=0.9.0
Pillow>=8.3.0
einops>=0.4.0
timm>=0.6.0